pyproject.toml
[project]
name = "hydrogen_pfhx"
version = "0.0.1"
description = "Chemical engineering model of hydrogen plate-fin heat exchanger"
readme = "README.md"
keywords = ['hydrogen','pfhx','heat exchanger','ortho','para','engineering']
authors = [
    { name = "Keelan O'Neill", email = "keelan.oneill@uwa.edu.au" }
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD-3-Clause"
]
dependencies = [
    "CoolProp",
    "docopt",
    "matplotlib",
    "msvc-runtime",
    "numpy",
    "pandas",
    "PyYAML",
    "scipy",
]
requires-python = ">=3.8"

[project.urls]
url = "https://github.com/fsr-uwa/hydrogen-pfhx"
